# [PERF] optimize updates to hasMany state