--- a/addon/-private/system/diff-array.js
+++ b/addon/-private/system/diff-array.js
@@ -2,6 +2,7 @@
   @namespace
   @method diffArray
   @private
+  @for DS
   @param {Array} oldArray the old array
   @param {Array} newArray the new array
   @return {hash} {