--- a/addon/-private/system/relationships/state/has-many.js
+++ b/addon/-private/system/relationships/state/has-many.js
@@ -5,6 +5,8 @@ import Relationship from './relationship';
 import OrderedSet from '../../ordered-set';
 import ManyArray from '../../many-array';
 
+import diffArray from '../../diff-array';
+
 export default class ManyRelationship extends Relationship {
   constructor(store, internalModel, inverseKey, relationshipMeta) {
     super(store, internalModel, inverseKey, relationshipMeta);
@@ -19,6 +21,7 @@ export default class ManyRelationship extends Relationship {
     this._loadingPromise = null;
     this._willUpdateManyArray = false;
     this._pendingManyArrayUpdates = null;
+    this._removedInternalModels = null;
   }
 
   _updateLoadingPromise(promise, content) {
@@ -203,6 +206,11 @@ export default class ManyRelationship extends Relationship {
     if (!this.members.has(internalModel)) {
       return;
     }
+    // remember removals so we can deal with them in computeChanges
+    if (!this._removedInternalModels) {
+      this._removedInternalModels = new OrderedSet();
+    }
+    this._removedInternalModels.add(internalModel);
     super.removeInternalModelFromOwn(internalModel, idx);
     // note that ensuring the many array is created, via `this.manyArray`
     // (instead of `this._manyArray`) is intentional.
@@ -266,22 +274,75 @@ export default class ManyRelationship extends Relationship {
   }
 
   computeChanges(internalModels = []) {
-    let members = this.canonicalMembers;
-    let internalModelsToRemove = [];
-    let internalModelSet = setForArray(internalModels);
+    let members = this.canonicalMembers.toArray();
+    let removedSinceLastTime = this._removedInternalModels;
+    this._removedInternalModels = null; // clear it for next time
 
-    members.forEach(member => {
-      if (internalModelSet.has(member)) { return; }
+    let diff = diffArray(members, internalModels);
 
-      internalModelsToRemove.push(member);
+    if (diff.firstChangeIndex === null) {
+      // no changes found
+      if (removedSinceLastTime) {
+        // records that have been removed since last compute will need their inverses to be corrected
+        for (let i = 0, l = internalModels.length; i < l; i++) {
+          let record = internalModels[i];
+          if (removedSinceLastTime.has(record)) {
+            this.flushCanonicalLater();
+            break;
+          }
+        }
+      }
+      return;
+    }
+    let removedMembersSet = setForArray(members.slice(diff.firstChangeIndex, diff.firstChangeIndex + diff.removedCount));
+    let changeBlockSet = setForArray(internalModels.slice(diff.firstChangeIndex, diff.firstChangeIndex + diff.addedCount));
+
+    // remove members that were removed but not re-added
+    removedMembersSet.forEach(member => {
+      if (!changeBlockSet.has(member)) {
+        this.removeCanonicalInternalModel(member);
+      }
     });
 
-    this.removeCanonicalInternalModels(internalModelsToRemove);
+    let flushCanonicalLater = false;
 
-    for (let i = 0, l = internalModels.length; i < l; i++) {
-      let internalModel = internalModels[i];
-      this.removeCanonicalInternalModel(internalModel);
-      this.addCanonicalInternalModel(internalModel, i);
+    // --- deal with records before the change block
+    if (removedSinceLastTime) {
+      // records that have been removed since last compute will need their inverses to be corrected
+      for (let i = 0; i < diff.firstChangeIndex; i++) {
+        let record = internalModels[i];
+        if (removedSinceLastTime.has(record)) {
+          flushCanonicalLater = true;
+          break;
+        }
+      }
+    }
+    // --- deal with records in the change block
+    for (let i = diff.firstChangeIndex, l = diff.firstChangeIndex + diff.addedCount; i < l; i++) {
+      let record = internalModels[i];
+      if (members[i] !== record) {
+        if (removedMembersSet.has(record)) {
+          // this is a reorder
+          this.removeCanonicalInternalModel(record);
+        }
+        // reorder or insert
+        this.addCanonicalInternalModel(record, i);
+      }
+    }
+    // --- deal with records after the change block
+    if (!flushCanonicalLater && removedSinceLastTime) {
+      // records that have been removed since last compute will need their inverses to be corrected
+      for (let i = diff.firstChangeIndex + diff.addedCount, l = internalModels.length; i < l; i++) {
+        let record = internalModels[i];
+        if (removedSinceLastTime.has(record)) {
+          flushCanonicalLater = true;
+          break;
+        }
+      }
+    }
+
+    if (flushCanonicalLater) {
+      this.flushCanonicalLater();
     }
   }