--- a/tests/unit/diff-array-test.js
+++ b/tests/unit/diff-array-test.js
@@ -1,5 +1,4 @@
 import { module, test } from 'qunit';
-
 import { diffArray } from 'ember-data/-private';
 
 module('unit/diff-array Diff Array tests', {