resolves #4150

On master:
With 5000 records, adding 1000 on the end with a `store.push` takes 250ms in `has-many.computeChanges`
Removing them again takes another 250ms

With this PR:
adding takes 37ms
removing takes 9ms