@BryanCrotaz 2017-03-21
thoughts?