@runspired 2018-04-19
@BryanCrotaz main thing would be to rebase and see where we are (do we now fail tests for instance). I am very happy to help you keep this moving forward so that it does not stall again.