@runspired 2018-04-22
I've rebased the commit, and I find that it happily fixes our `ASSERT KNOWN LIMITATION EXISTS` test for hasMany losing local changes after a delete.  That said, I'm not sure this approach will always cause us to keep local changes, and what's more I suspect this leaks `_removedInternalModels`.

I'm going to add some more tests and investigate those two points.